+++ tests/testthat/test-trace.R
@@ -10,9 +10,9 @@ test_that("safe_untrace makes recording not error if not already traced", {
   expect_error(stop_capturing(), NA)
 })
 
-test_that("quietly muffles messages, conditional on httptest.debug", {
+test_that("quietly muffles messages, conditional on httptest2.debug.trace", {
   expect_message(quietly(message("A message!")), NA)
-  options(httptest.debug = TRUE)
-  on.exit(options(httptest.debug = NULL))
+  options(httptest2.debug.trace = TRUE)
+  on.exit(options(httptest2.debug.trace = NULL))
   expect_message(quietly(message("A message!")), "A message!")
 })