+++ NEWS.md
@@ -6,6 +6,7 @@ Initial port of `httptest` to work on top of `httr2`. All APIs preserved except:
 * Redacting functions are applied to both the `httr2_response` and `httr2_request` objects, the latter in order to be able to alter the URL/mock file path being written. You do not need separate "redactors" and "requesters".
 * `expect_header()` has been renamed to `expect_request_header()` and takes headers to match by argument name rather than a single string "Header: value". This allows more flexibility in matching headers, plus the ability to match multiple headers. It also no longer uses `warning()` to extract the headers, so test failures are cleaner.
 * The failure message when an unexpected request is made is more clear and is distinguished from other "errors". Expected mock file paths are more clearly shown when appropriate, and they are no longer included in the message itself that you would assert in `expect_GET()`, `expect_POST()`, et al.
+* All `options()` have been renamed to match the package name (e.g. `options(httptest2.verbose)`)
 
 Changes to function signatures:
 
@@ -20,6 +21,4 @@ Some functions were removed:
 * `public()`
 * `fake_response()` (just use `httr2::response()`)
 * `gsub_request()` and `set_requester()` (no longer necessary)
-
-Other improvements: